shopify-data-ingestion/requirements.txt
google-cloud-bigquery
google-auth
requests